Made a 'FloatingWindow' component that the CodeSnippets and GUIMode both use
